myproject/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
bcrypt==4.1.2
boto3==1.34.80
botocore==1.34.80
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.6
jmespath==1.0.1
multidict==6.0.5
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
twilio==9.0.4
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
